Move data classes to package model. To keep the rule
